:recycle: (canvas) fixing codeclimate errors
