+++ src/lib/particles/.eslintrc.js
@@ -1,0 +1,6 @@
+
+module.exports = {
+  rules: {
+    'no-param-reassign': 'off',
+  },
+};
